Remark: mismatch between top/left and width/height of shapes
I stated in a previous commit that the top, left, width and height
properties of the shapes to plot referred to the bounding box of
their parent cell; actually, only the top and left property refer
to the bounding box of the parent cell; the width and height property
refer to the shape itself.

In order to resolve this mismatch, I will compute the actual top
and left values of the shape taking the vertical and horizontal
alignments into account. I will then remove the alignment properties
from the shape.

I will first need to retrieve the width and height of the parent cell;
another way to resolve the mismatch is to set the box properties to
cellTop, cellLeft, cellWidth and cellHeight properties.