Remote intermediate properties from shape object
I removed properties related to the parent and alignement within
the parents, used as intermediate values to compute the absolute
position of the shape in tiles within the chart.

Note that in center and middle alignments, the position value
can contain an integer number of tiles plus an half tile (0.5).
These values should be rounded to the nearest integer up or down
to preserve alignment of tiles with the grid.