Create two separate functions to add tiles in the odd and even sequences
At this point, both functions have the same implementation, copied as is
from the function addTileAndSymmetricTiles() previously used for both the
odd and the even sequences.

The goal of this ongoing refactoring is to remove the gap between the
four quarters of the circles drawn with the even sequence.

The circle sector shall no longer be shifted for the odd circle, and the
computation of positions for symmetric points shall be customized.