Fix computation of symmetrical positions in odd circle sequence
In this updated computation, x and y are replaced with xSym = x + 1
and ySym = y + 1 in some places. The list of points inserted at each
step becomes:

1: x, y
2: -xSym, -ySym
3: -xSym, ySym
4: x, -ySym
5: -ySym, x
6: y, -xSym
7: -ySym, -xSym
8: y, x

Also, I removed the checks for tiles on the vertical and horizontal axes
of symmetry, as well as their intersection point, which never occurs
in the odd circle sequence: tiles are added on each side of these axes,
never on the axes themselves, which lie at the edge of tiles.