Add SVG export
This is a simple copy of the HTML generated by d3.js, using the
innerHTML property. The exported SVG lacks the SVG namespace as
well as indentation. This could be fixed by traversing the SVG
node trees and implementing a custom serialization, or by
generating the expected SVG as text in the first place. This
generated SVG text could then be used to create the initial
charts instead of using the d3.js API.