Rename setLayoutDimensions() to setTableLayoutDimensions()
I also introduced an intermediate method, called setLayoutDimensions().
This method just calls setTableLayoutDimensions() at this point. I will
then add triage code to apply the table layout processing only to
charts and table layouts, while processing charts as a list:

A table layout has the following structure:

[
  ["table-layout",...], // header row with column margins
  [row-header,...],  // row header cell for height followed with data cells
  ...
]

A chart has the same structure, with a single row and column
and the text "chart" instead of "table-layout" in top-left cell.

The structure of a list of charts is:
[
  ["charts", ...], // header with list of chart names
  [
    // list of charts
  ]
]

A chart has a single row besides the header, which contains
a list of "chart" structures.