Switch parameters name and value in callback of forEachProperty
I noticed at first use that the callback may need the value
without the name. The contrary seems less likely. I thus decided
to make the definition of the name optional in the callback,
pushing the name in last position.

This change is consistent with the definition of the callback
in the utility function forEach(), which also provides the value
as first argument to the callback, and the index, which is also
the name of a property in an array, as second argument.