Add module to collect filters
The module listens to the event "filter-selected"
and keeps track of the set of filters currently selected.
To notify other modules of changes in the set of filters,
it publishes the events "filter-created" when a filter which
was not selected gets selected, "filter-updated" when a filter
gets selected with a new value (no check is performed that the
value is actually different), and "filter-deleted" when a
filter which was previously selected is reset to an inactive
state (i.e. matches all authors, thus filtering nothing).