Filter authors using the combined selector
At this step the selector is applied to all contributions
of every single author. At a later step, it will only be
applied to the set of authors matching all criteria separately
(which is a superset of the expected result for funnel filtering).

I noticed multiple authors displayed when I select the filter
for 18 contributions, while the corresponding category contains
a single author. I will investigate the reason of the discrepancy.