Add missing initialization of minimum width/height of table columns/rows
The headers were the width and height of columns/rows are computed contain
a text in the initial layout descriptions. The comparisons between these
strings and actual numbers using the operator > in max() resulted in false,
and the number value was selected to replace the string description. However,
the opposite comparison using the operator < would also result in false,
and the text description would be selected instead of the expected number
description. In order to avoid the error, which happens when the custom
implementation of max() is replaced with standard Math.max() in utils.js,
I defined a new function which discards the text value no matter what.