Prepare to update the counts displayed in each filter option
This is an intermediate step towards the display of dynamic values
in each option of filters: first the total of authors of the category
currently selected, next the total of authors in the category.

Two separate functions, one to fill the list initially, one to
update the list when the set of selected authors changes, will
share a set of utility functions to prepare the display.

Another feature expected is the filtering of options displayed
to remove options which match no author in current selection.