Generalize: list all predicted values for TOTAL_CONTRIBUTIONS too
Before, the list of predicted value for this filter was special:
it was a number which indicated the maximum number of contributions
which would match for an author. I updated this property to list
all allowed values as strings, like in the categories data, from
"1" to the maximum allowed value.