Fix the total of selected authors displayed in default item
I added an try with the name "" in predictive filters,
which counts the number of authors matching the relaxed
filter with any value. This number is consistent with
the information displayed for other filters: this is
the actual number of authors selected when the default item
is selected based on the current set of active filters.