Use system fonts in select lists as fall back
As of today, Safari and IE do not support webfonts in select lists.
Since we need a monospace font to preserve the alignment of padded
information in the filter lists, I applied system fonts to the
select lists instead.

I listed the web font last, with the intent of using it instead of
system fonts on operating systems without Safari or IE anyway.