Wrap charts in a new "charts" structure in "layout" event
In the next step, I will sort and filter the list of charts,
which requires to update the list in the original structure,
or wrap the new list in a new structure.

I chose the second option, which results in a single chart
being wrapped in a "charts" structure as well.