Bug fix: reset selection in total contributions filter
The default item has value "1", not "" in total contributions
filter. Note that for the same reason, the code in
getTotalAuthorsInCategory() will read the total number
of authors from the categories data rather than using the
short cut for the total contributions filter. This is
not a problem, however, since the default item is
present in the set of categories for total contributions,
which is not the case for other filters.