Add rect elements for the bounding box of each tile
While the circle element and the surrounding group only extended
to the surface of the circle, the bounding box rect covers the
whole tile, which will facilitate the selection of a tile to obtain
extra information about an author. Note that the parent g element now
covers the whole tile as well.