Wrap charts in a div to apply extra CSS styles (min-height)
The min-height applied to the SVG element with a "height"
attribute had no effect. I first tried to apply a minimum
height value, but this was not flexible enough since it
was applied to all types of charts. I then thought of
wrapping each chart in an extra div, with the class 'chart-box'
for all charts, and I moved the CSS class which reflects
the grouping from the SVG element to the parent div.