Highlight author circles on hover
The circle is highlighted when the mouse cursor hovers either
over the tile circle or the adjacent tile bounding box.

The hover effect is a bit slow in Firefox, not as responsive
as in Safari. I will check whether applying the effect with
a descendent selector when a grouping tile is hovered makes
any difference.