Handle categories not selected in getAuthorsSelectedInGroup()
An empty list is now correctly returned in this case, which
used to result in a failure. This issue was introduced when
converting the previous algorithm, which computed the number
of authors without keeping the list, missing the fact that
a total of zero authors was returned by default when the
category was not found.